feat: enhance TransformerBlock and MiniGPT with RoPE support and positional encoding options

+++ llm_lab/core/model/mlp.py
@@ -7,7 +7,7 @@ import torch
 class FeedForwardConfig:
     d_model: int
     d_ff: int
-    dropout: int 
+    dropout: float 
 
 class FeedForward(nn.Module):
     def __init__(self,config :FeedForwardConfig):
